requirements.txt
Flask
supabase
google-generativeai
Pillow
python-dotenv
requests
